email_categorisation/requirements.txt
SQLAlchemy~=2.0.31
APScheduler~=3.10.4
fastapi~=0.111.1
tomlkit~=0.13.2
uvicorn~=0.30.3
pydantic~=2.8.2
pydantic-settings~=2.4.0
sentry-sdk~=2.14.0
gunicorn~=23.0.0
loguru~=0.7.2
PyJWT~=2.3.0
python-dotenv~=1.0.1
prettytable~=3.11.0